fix: split denylist scanner
